wallet drawer user projects query fn
